map-service/map.requirements.txt
flask
flask-cors
azure-cognitiveservices-speech